+++ app/src/main/java/com/the_tinkering/wk/Constants.java
@@ -48,7 +48,7 @@ public final class Constants {
     /**
      * The number of theme customization options offered.
      */
-    public static final int NUM_THEME_CUSTOMIZATION_OPTIONS = 26;
+    public static final int NUM_THEME_CUSTOMIZATION_OPTIONS = 32;
 
     /**
      * Short names of weekdays for timeline bar chart.